+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### Changed
 * Moved color models to their own sub-modules
-* composite methods only available for premultiplies Rasters
+* composite methods only available for premultiplied Rasters
 
 ## [0.11.0] - 2020-04-11
 ### Added